+++ R/naturecounts-pkg.R
@@ -1,0 +1,14 @@
+#' Access and download data on plant and animal populations from NatureCounts
+#' 
+#' naturecounts is an R package for accessing and downloading data on plant and
+#' animal populations from various databases through
+#' [NatureCounts](https://www.naturecounts.ca/), a service managed by Bird
+#' Studies Canada.
+#' 
+#' See the vignettes (`vignettes(package = "naturecounts")`, or at 
+#' https://birdscanada.github.io/naturecounts for getting started!
+#' 
+#' @importFrom memoise memoise
+#' @importFrom dbplyr sql
+"_PACKAGE"
+