+++ tests/testthat/test_01_server.R
@@ -1,5 +1,3 @@
-context("Server Calls")
-
 test_that("srv_query() returns data", {
 
   expect_silent(d <- srv_query(api$bmde_versions))