+++ tests/testthat/test_05_utils.R
@@ -1,5 +1,3 @@
-context("Utility functions")
-
 test_that("as_numeric() converts to numeric if possible", {
   expect_is(as_numeric(13), "numeric")
   expect_is(as_numeric("13"), "numeric")
@@ -27,4 +25,18 @@ test_that("capture_df() capture dataframe", {
     expect_is("character") %>%
     expect_length(1) %>%
     expect_match("[a-zA-Z ]+$")
+})
+
+test_that("nc_remove_cache()", {
+  expect_true(memoise::is.memoised(srv_query))
+  
+  p <- srv_query(path = api$permissions, token = srv_auth("sample"))
+  
+  expect_true(memoise::has_cache(srv_query)(path = api$permissions, 
+                                            token = srv_auth("sample")))
+
+  expect_true(nc_remove_cache())
+  expect_false(memoise::has_cache(srv_query)(path = api$permissions, 
+                                             token = srv_auth("sample")))
+  
 })